create solar choukei module
